requirements.txt
streamlit==1.44.1
pandas==2.2.3
python-docx==1.1.2
pytest==8.3.5
